fixed more smooth scroll-to-top button
