+++ GiEnJul/ClientApp/src/components/landingPage/Macro.tsx
@@ -1,7 +1,7 @@
-import { Button, Container, Grid, Typography } from "@material-ui/core";
+import { Button, Container, Grid, Typography, useScrollTrigger, Fab } from "@material-ui/core";
 import React, { useCallback } from "react";
 import { useHistory } from "react-router-dom";
-import Tab from "../../common/components/Tab";
+import ScrollToTop from "../../common/components/ScrollToTop";
 import Companies from "./Companies";
 import How from "./How";
 import Questions from "./Questions";
@@ -9,7 +9,10 @@ import useStyles from "./Styles";
 import logo from './../../styling/img/logo_gronn.svg';
 import Contact from "./Contact";
 import NavMenu from "../../common/components/NavMenu";
-import {ArrowBackIos, ArrowForwardIos} from '@material-ui/icons';
+import {ArrowForwardIos} from '@material-ui/icons';
+
+
+
 
 const Home = () => {
   const classes = useStyles();
@@ -54,17 +57,7 @@ const Home = () => {
       <Questions />
       <Companies />
       <Contact/>
-      <Grid item xs={1}>
-        <Tab maxPagePosition={300} path="top" />
-      </Grid>
-      <Grid
-        container
-        direction="row"
-        justifyContent="center"
-        alignItems="center"
-      >
-        {/* <Tab maxPagePosition={140} path="/bli-giver" /> */}
-      </Grid>
+        <ScrollToTop maxPagePosition={300}/>
     </Container>
     </>
   );
